// generated/src/app/api-models/ipa-provider.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IpaProvider {
  ipaId: string;
  ipaProvCategory: string;
  effectiveDate: Date;
  seqProvId: number;
  termDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  contractInHouseFlag: string;
}